Finished Ch 3.1 Leftist Heaps exercises 3.3 and 3.4.
The trick was to accept no benefits or clarity would be achieved from
completing this particular exercise, but instead a path/template is created from
which benefits in lazy/concurrent environments (exercise 3.4) are
obvious.

I have not tested the solution to the exercise, but it seems to make
sense. Moving on.